+++ relay-server/src/endpoints/common.rs
@@ -598,6 +598,7 @@ where
         .send(Create {
             scoping,
             length: None,
+            attachment_type: item.attachment_type(),
         })
         .await
         .ok()?